msm: thermal: Pass correct size of voltage table to IOCTL
Fix issue in msm_thermal_process_voltage_table_request.
For voltage tables bigger than 16 it is not correclty
passing the partial table size to user space.
The full size is passed instead of the partial size.

For example if reading a table of 18 values the
first read returns 16, and the partial read returns 18
but it should return 2.

Change-Id: I75943e94341388cca772ee45bc1275fb5d2091d2
Signed-off-by: Jeff Bernard <jbernard@codeaurora.org>
Signed-off-by: Santhosh Kumar Thimmanna Bhattar <sthim@codeaurora.org>
Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>